# services/api_vbox/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "API vbox",
    "description": "OpenAPI of API vbox",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "https://localhost:8000/",
      "description": "API dev server"
    }
  ],
  "paths": {
    "/list": {
      "get": {
        "tags": [
          "list"
        ],
        "summary": "Vbox Manage List",
        "description": "Apply `VBoxManage list <directive>` with parameter `q` a list of directives.",
        "operationId": "vbox_manage_list_list_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Q",
              "maxLength": 98,
              "minLength": 3,
              "pattern": "^((vms|runningvms|intnets|hostinfo|groups|systemproperties|bridgedifs|hostonlyifs|natnets|dhcpservers),?)+$",
              "type": "string"
            },
            "name": "q",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Sort",
              "type": "boolean",
              "default": false
            },
            "name": "sort",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Long",
              "type": "boolean",
              "default": false
            },
            "name": "long",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/import": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Vbox Manage Import",
        "description": "Apply `VBoxManage import` command depending on {ovf_params}.\n\nArgs:\n    request (Request): request emitted\n    ovf_params (OvfSessionParams): ovf parameters\n\nRaises:\n    HTTPException: ovf path {ovf_params.image} does not exist\n    HTTPException: error threw during the execution of `VBoxManage import`\n\nReturns:\n    BasicResponse: success response",
        "operationId": "vbox_manage_import_import_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OvfSessionParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/startvm": {
      "post": {
        "tags": [
          "startvm"
        ],
        "summary": "Vbox Manage Startvm",
        "description": "Apply `VBoxManage startvm` command depending on {vm_params}.\n\nIt will start a VM.\n\nArgs:\n    request (Request): request emitted\n    vm_params (StartVMParams): startvm directive parameters\n\nRaises:\n    HTTPException: vm with given name {vmname} does not exist\n    HTTPException: error threw during the execution of `VBoxManage startvm`\n\nReturns:\n    BasicResponse: success response",
        "operationId": "vbox_manage_startvm_startvm_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartVMParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BasicError": {
        "title": "BasicError",
        "required": [
          "error",
          "code"
        ],
        "type": "object",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "title": "Details"
          }
        },
        "description": "API error response model\n\nerror: message to describe a error\ncode: code error\ndetails: details error"
      },
      "BasicResponse": {
        "title": "BasicResponse",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "items": {
            "title": "Items",
            "type": "object",
            "default": {}
          }
        },
        "description": "API response model\n\nmsg: message to describe a response\nitems: data to submit with the response"
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "OvfSessionParams": {
        "title": "OvfSessionParams",
        "required": [
          "image",
          "vmname"
        ],
        "type": "object",
        "properties": {
          "image": {
            "title": "Image",
            "type": "string",
            "format": "path"
          },
          "vmname": {
            "title": "Vmname",
            "type": "string"
          },
          "options": {
            "title": "Options",
            "type": "array",
            "items": {
              "type": "boolean"
            },
            "default": []
          },
          "vsys": {
            "title": "Vsys",
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "default": 0
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer",
            "default": 0
          }
        },
        "description": "Datamodel for OVF session parameters.\n\nAttributes:\n    image (Path[str]): specify ovf image path\n    vmname (str): name to assign to the vm\n    options (List[bool]): options to keep in the following order\n                          between 'keepallmacs', 'keepnatmacs', 'importtovdi'\n    vsys (Union[int, str]): virtual system to use\n    expires_in (Optional[int]): time to live in seconds of the vm. Specify when to delete it."
      },
      "StartVMParams": {
        "title": "StartVMParams",
        "required": [
          "vmname"
        ],
        "type": "object",
        "properties": {
          "vmname": {
            "title": "Vmname",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string",
            "default": "headless"
          }
        },
        "description": "Datamodel for VBoxManage StartVM directive\n\nAttributes:\n    vmname (str): VM's name to start\n    type (Optional[str]): type of start [\"headless\" | \"gui\" | \"separate\"]"
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "apiAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "apiAuth": []
    }
  ]
}